refactor(mobile): extract submit button
